+++ src/components/loading/docs/index.js
@@ -1,0 +1,7 @@
+import Markdown from '../../../../libs/markdown';
+
+export default class Loading extends Markdown {
+    document() {
+        return require(`./Loading.md`);
+    }
+}