+++ src/index.js
@@ -22,6 +22,7 @@ export { default as StepBar } from './components/step-bar';
 export { default as Table } from './components/table';
 export { default as Tag } from './components/tag';
 export { default as Notification } from './components/notification';
+export { default as Loading } from './components/loading';
 export { PropTypes } from '../libs';
 export { Transition } from '../libs';
 export { View } from '../libs';